added environments and bug fixes

+++ lib/pages/new_pricing_profile_page/NewPricingProfilePageReducer.dart
@@ -27,7 +27,7 @@ NewPricingProfilePageState _updateDepositAmount(NewPricingProfilePageState previ
   resultCost = resultCost.replaceAll(' ', '');
   double doubleCost = resultCost.isNotEmpty ? double.parse(resultCost) : 0.0;
   return previousState.copyWith(
-    deposit: doubleCost,
+    deposit: doubleCost.toInt(),
   );
 }
